refactor : Lottos, Lotto 패키지 이동 (vo -> model)
