+++ src/main/java/lotto/model/LottoGenerator.java
@@ -1,7 +1,6 @@
 package lotto.model;
 
 import java.util.List;
-import lotto.vo.Lotto;
 
 public class LottoGenerator {
