+++ src/main/java/lotto/model/Lottos.java
@@ -1,9 +1,10 @@
-package lotto.vo;
+package lotto.model;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import lotto.model.Lotto;
 
 public class Lottos {
