+++ src/test/java/lotto/model/LottoGeneratorTest.java
@@ -1,10 +1,8 @@
 package lotto.model;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
-import lotto.vo.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
