+++ src/test/java/lotto/vo/LottoTest.java
@@ -3,6 +3,7 @@ package lotto.vo;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Arrays;
+import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
